Moved classes from wireviz.py to new files
